Entity 설계
