+++ src/main/java/com/depromeet/breadmapbackend/breadShops/domain/Menus.java
@@ -1,0 +1,10 @@
+package com.depromeet.breadmapbackend.breadShops.domain;
+
+import lombok.Getter;
+
+import javax.persistence.Entity;
+
+@Entity
+@Getter
+public class Menus {
+}
