+++ src/main/java/com/depromeet/breadmapbackend/flags/domain/Flags.java
@@ -1,0 +1,10 @@
+package com.depromeet.breadmapbackend.flags.domain;
+
+import lombok.Getter;
+
+import javax.persistence.Entity;
+
+@Entity
+@Getter
+public class Flags {
+}
