+++ src/main/java/com/depromeet/breadmapbackend/members/domain/Members.java
@@ -1,8 +1,10 @@
-package com.depromeet.breadmapbackend.domain.members;
+package com.depromeet.breadmapbackend.members.domain;
 
-import com.depromeet.breadmapbackend.domain.BaseEntity;
-import com.depromeet.breadmapbackend.domain.breadShops.BreadShops;
+import com.depromeet.breadmapbackend.common.domain.BaseEntity;
+import com.depromeet.breadmapbackend.breadShops.domain.BreadShops;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
@@ -13,6 +15,8 @@ import static javax.persistence.GenerationType.IDENTITY;
 
 @Getter
 @Entity
+@NoArgsConstructor
+@AllArgsConstructor
 public class Members extends BaseEntity {
 
     @Id
@@ -20,11 +24,17 @@ public class Members extends BaseEntity {
     @Column(name = "member_id") // 임의로 이름 정했습니다!
     private Long id;
 
-    @OneToMany(mappedBy = "pioneer")
+    @OneToMany(mappedBy = "members")
     private List<BreadShops> breadShopsList = new ArrayList<>();
 
+    @Column(nullable = false)
     private String name;
+
+    @Column(nullable = false)
     private String encryptedPassword;
+
+    @Column(nullable = false)
     private String email;
+
     private Integer breadTestResult;
 }
